Share history test setup
